Review pattern for img-paths if `post_asset_folder = true`
`[\\{0,1}\/{0,1}]` matches any character of `\`, `{`, `0`, `,`, `1`, `}`, '/`.
`[^\/^\\]´ matches any character not of `/`, `^`, `\\`

I'm not sure what these regex should match exactly, but this seems not intentional and I think `(ret[0].length == s.length` is never `true`. If I change it to something I could think of is the intention, `(ret[0].length == s.length` would be `true`, but the paths are  not correct anymore.
Thus I removed the whole block since it's never executed (I think).